+++ docs/CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
   Docker image
 * Refactor ``amplicon-report`` process to contain all relevant inputs for
   ``amplicon-archive-multi-report``
+* Refactor ``amplicon-archive-multi-report``
 
 
 ==================